[Stratum] Add `-defaultminingaddress` for wallet-less mining.
Default credentials are a convenient way to let the server manage mining rewards, so that individual miners don't have to all be individually configured.  This provides an alternative to `-stratumwallet`, which generates new addresses for each block but requires wallet support to be compiled in and a wallet loaded at runtime.  By using `-defaultminingaddress` the server operator can have all block rewards go to a specified address when default credentials are used.